Extract counting XML event filter to separate class
