+++ intellij-plugin-structure/structure-ide/src/main/java/com/jetbrains/plugin/structure/xml/XmlStreamings.kt
@@ -1,8 +1,5 @@
 package com.jetbrains.plugin.structure.xml
 
-import com.jetbrains.plugin.structure.ide.dependencies.PluginXmlDependencyFilter
-import org.slf4j.Logger
-import org.slf4j.LoggerFactory
 import java.io.Closeable
 import java.io.InputStream
 import java.io.OutputStream
@@ -14,7 +11,6 @@ import javax.xml.stream.XMLOutputFactory
 import javax.xml.stream.XMLStreamException
 import javax.xml.stream.XMLStreamReader
 import javax.xml.stream.XMLStreamWriter
-import javax.xml.stream.events.XMLEvent
 
 fun XMLInputFactory.newEventReader(inputStream: InputStream): CloseableXmlEventReader =
   CloseableXmlEventReader(createXMLEventReader(inputStream))
@@ -36,39 +32,6 @@ class CloseableXmlEventWriter(private val delegate: XMLEventWriter) : XMLEventWr
   @Throws(XMLStreamException::class)
   override fun close() {
     delegate.close()
-  }
-}
-
-class CountingXmlEventWriter(private val delegate: XMLEventWriter) : XMLEventWriter by delegate, Closeable {
-  private val eventCounter = hashMapOf<XmlEventType, Int>()
-
-  override fun add(event: XMLEvent) {
-    delegate.add(event)
-    val type = event.eventType
-    eventCounter[type] = (eventCounter[type] ?: 0) + 1
-  }
-
-  @Throws(XMLStreamException::class)
-  override fun close() {
-    if ((eventCounter.size == 1 && eventCounter[XMLEvent.START_DOCUMENT] == 1) || eventCounter.isEmpty()) {
-      // closing without an actual document being written
-      try {
-        delegate.close()
-      } catch (e: Exception) {
-        when (e) {
-          is RuntimeException, is XMLStreamException -> {
-            LOG.atError().log("Failed to close delegate XML event writer: {}", e.message)
-            return
-          }
-        }
-      }
-    } else {
-      delegate.close()
-    }
-  }
-
-  companion object {
-    private val LOG: Logger = LoggerFactory.getLogger(PluginXmlDependencyFilter::class.java)
   }
 }
